+++ docu/sidebars.js
@@ -8,12 +8,12 @@ module.exports = {
     {
       type: 'category',
       label: 'User Guide',
-      items: ['use/variant', 'use/grouping', 'use/matching', 'use/filter', 'use/recursive', 'use/generic', 'use/redux', 'cheat'],
+      items: ['use/variant', 'use/grouping', 'use/matching', 'use/filter', 'use/advanced-creation', 'use/redux', 'cheat'],
     },
     {
       type: 'category',
       label: 'Articles',
-      items: ['articles/that-type', 'articles/type-name', 'articles/augment'],
+      items: ['articles/that-type', 'articles/type-name', 'articles/augment', 'changelog'],
     },
     'api',
   ],